ONNX&YOLO Converter (#185)
* The parser has been tested on the resnest101e_Opset16.onnx,
GoogLeNet.onnx, and densenet121_Opset16.onnx models.
- The yolo11x-cls.pt model has been converted to the onnx format using
ultralytics.
- The parser for our first model has been fixed to use a single json
file style and a single code for reading and converting to a tensor.
* I removed the check for the equality of the bias to the last dimension
of the weights from the tensor implementation, as in the onnx and yolo
models, the bias is equal in size to the first array.
- Additionally, the weights and bias data are now stored more explicitly
in json, with each data item in its own fields "weights": and "bias":,
which helps prevent potential errors during conversion.
* I also think that we should provide a link to download each model, as
we won't be able to store them all in the repository. We should also
configure the "evaluate model" check in github to download the model
from the link and enable caching.